# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentVoice API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  title: string;
  content: string;
  category: string;
  likeCount: number;
  commentCount: number;
  createdAt: string;
}

export interface Comment {
  id: number;
  postId: number;
  content: string;
  createdAt: string;
}

export type PostWithComments = Post & {
  comments: Comment[];
};

export interface CreatePostInput {
  title: string;
  content: string;
  category: string;
}

export interface CreateCommentInput {
  content: string;
}

export type AdminStatsPostsByCategoryItem = {
  category: string;
  count: number;
};

export interface AdminStats {
  totalPosts: number;
  postsByCategory: AdminStatsPostsByCategoryItem[];
  mostLikedPosts: Post[];
  recentPosts: Post[];
}

export type ListPostsParams = {
  category?: string;
  sort?: ListPostsSort;
};

export type ListPostsSort = (typeof ListPostsSort)[keyof typeof ListPostsSort];

export const ListPostsSort = {
  latest: "latest",
  most_liked: "most_liked",
} as const;