# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudentVoice API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Post operations
  - name: comments
    description: Comment operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, most_liked]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostWithComments"
        "404":
          description: Not found

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated like count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /posts/{id}/comments:
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [posts]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - category
        - likeCount
        - commentCount
        - createdAt

    PostWithComments:
      allOf:
        - $ref: "#/components/schemas/Post"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
          required:
            - comments

    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - content
        - createdAt

    CreatePostInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
      required:
        - title
        - content
        - category

    CreateCommentInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    AdminStats:
      type: object
      properties:
        totalPosts:
          type: integer
        postsByCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required:
              - category
              - count
        mostLikedPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        recentPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required:
        - totalPosts
        - postsByCategory
        - mostLikedPosts
        - recentPosts