# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentVoice API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsPostsByCategoryItem } from "./adminStatsPostsByCategoryItem";
import type { Post } from "./post";

export interface AdminStats {
  totalPosts: number;
  postsByCategory: AdminStatsPostsByCategoryItem[];
  mostLikedPosts: Post[];
  recentPosts: Post[];
}