# lib/api-zod/src/generated/types/createPostInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentVoice API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostInput {
  title: string;
  content: string;
  category: string;
}