# lib/api-zod/src/generated/types/postWithComments.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentVoice API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from "./comment";
import type { Post } from "./post";

export type PostWithComments = Post & {
  comments: Comment[];
};